consolidate into into a single job
