+++ scripts/bg_manage/filter_namespaces.py
