+++ scripts/build_env/main.py
@@ -2,12 +2,12 @@ from envgenehelper import *
 from envgenehelper.deployer import *
 
 from build_env import build_env, process_additional_template_parameters
-from cloud_passport import update_env_definition_with_cloud_name
+from scripts.cloud_passport.cloud_passport import update_env_definition_with_cloud_name
 from create_credentials import create_credentials
 from render_config_env import EnvGenerator
 from resource_profiles import get_env_specific_resource_profiles
 
-from filter_namespaces import apply_ns_build_filter
+from scripts.bg_manage.filter_namespaces import apply_ns_build_filter
 
 
 INVENTORY_DIR_NAME = "Inventory"
