+++ scripts/sd/process_sd.py
@@ -19,6 +19,8 @@ from envgenehelper.sd_helper import (basic_merge_multiple, MergeType, calculate_
                                      SD_FILE_NAME, DELTA_SD_FILE_NAME)
 from typing_extensions import deprecated
 
+from scripts.pipeline.orchestrator import PipelineParametersHandler
+
 MERGE_METHODS = {
     MergeType.BASIC: helper.basic_merge,
     MergeType.BASIC_EXCLUSION: helper.basic_exclusion_merge,
@@ -66,28 +68,6 @@ def handle_deploy_postfix_namespace_transformation(sd_data: dict, namespace_dict
     return sd_data
 
 
-def prepare_vars_and_run_sd_handling(
-    base_dir: str = None,
-    env_name: str = None,
-    cluster: str = None,
-    sd_source_type: str = None,
-    sd_version: str = None,
-    sd_data: str = None,
-    sd_delta: str = None,
-    sd_merge_mode: str = None,
-    operation_type: OperationType = None,
-):
-    base_dir = base_dir or getenv_and_log('CI_PROJECT_DIR')
-    env_name = env_name or getenv_and_log('ENVIRONMENT_NAME')
-    cluster = cluster or getenv_and_log('CLUSTER_NAME')
-    sd_source_type = sd_source_type or getenv('SD_SOURCE_TYPE')
-    sd_version = sd_version or getenv('SD_VERSION')
-    sd_data = sd_data or getenv('SD_DATA')
-    sd_delta = sd_delta or getenv('SD_DELTA')
-    sd_merge_mode = sd_merge_mode or getenv("SD_REPO_MERGE_MODE")
-    operation_type = operation_type or OperationType(getenv("OPERATION_TYPE"))
-
-    env = Environment(base_dir, cluster, env_name)
     handle_sd(env, sd_source_type, sd_version, sd_data, sd_delta, sd_merge_mode, operation_type)
 
 
@@ -161,8 +141,17 @@ def multiply_sds_to_single(sds_data, effective_merge_mode):
     return full_sd_from_pipe
 
 
-def handle_sd(env, sd_source_type, sd_version, sd_data, sd_delta, sd_merge_mode, operation_type: OperationType):
+def handle_sd(handler: PipelineParametersHandler):
+    sd_source_type = handler.params.get('SD_SOURCE_TYPE')
+    sd_version = handler.params.get('SD_VERSION')
+    sd_data = handler.params.get('SD_DATA')
+    sd_delta = handler.params.get('SD_DELTA')
+    sd_merge_mode = handler.params.get("SD_REPO_MERGE_MODE")
+    operation_type = OperationType(handler.params.get("OPERATION_TYPE"))
+
+    env = Environment(handler.base_dir, handler.cluster, handler.env_name)
     base_sd_path = Path(f'{env.env_path}/Inventory/solution-descriptor/')
+
     if operation_type == OperationType.DEPLOY:
         namespace_names = getenv("NAMESPACE_NAMES")
         if namespace_names:
@@ -344,7 +333,3 @@ def get_appdef_for_app(appver: str, app_name: str, plugins: PluginEngine) -> art
     app_dict['registry'] = artifact_models.parse_registry(helper.openYaml(reg_def_path))
     app_def = artifact_models.Application.model_validate(app_dict)
     return app_def
-
-
-if __name__ == "__main__":
-    prepare_vars_and_run_sd_handling()
