+++ scripts/tests/app_reg_defs/test_appregdefs.py
