+++ scripts/tests/env_inventory_generation/test_env_inv_generation.py
