+++ scripts/tests/namespace_filter/test_namespace_filter.py
