+++ scripts/tests/sd/test_process_sd_artifact.py
@@ -6,7 +6,7 @@ import pytest
 from envgenehelper.env_helper import Environment
 from envgenehelper.models import OperationType
 
-from scripts.build_env.tests.base_test import BaseTest
+from scripts.tests.base_test import BaseTest
 from test_sd_helpers import do_prerequisites, assert_sd_contents, load_test_pipeline_sd_data
 
 os.environ['ENVIRONMENT_NAME'] = "temporary"
