+++ scripts/tests/sd/test_process_sd_local.py
@@ -4,7 +4,7 @@ import pytest
 from envgenehelper import logger, SD_FILE_NAME, OperationType
 from envgenehelper.env_helper import Environment
 
-from scripts.build_env.tests.base_test import BaseTest
+from scripts.tests.base_test import BaseTest
 from test_sd_helpers import do_prerequisites, assert_sd_contents, load_test_pipeline_sd_data
 
 os.environ['ENVIRONMENT_NAME'] = "temporary"
