+++ scripts/tests/sd/test_sd_helpers.py
