+++ scripts/utils/schema_validation.py
